+++ hmsc/updaters/updateBetaLambda.py
@@ -4,7 +4,7 @@ from hmsc.utils.tf_named_func import tf_named_func
 tfm, tfla, tfr, tfs = tf.math, tf.linalg, tf.random, tf.sparse
 
 @tf_named_func("betaLambda")
-def updateBetaLambda(params, data, priorHyperparams, dtype=np.float64):
+def updateBetaLambda(params, data, priorHyperparams, dtype=np.float32):
     """Update conditional updater(s):
     Beta - species niches, and
     Lambda - species loadings.
